* rename Slider, add orientation and add documentation + add some documentation + add some observables
